Revert "refactor: apply SOLID concepts to home screen"
