+++ src/screens/HomeScreen.tsx
@@ -1,0 +1,243 @@
+import React, { useEffect, useState } from 'react';
+import styled from 'styled-components/native';
+import { FlatList, RefreshControl, TouchableOpacity } from 'react-native';
+import { Button, Icon } from 'react-native-elements';
+import { FontAwesome } from '@expo/vector-icons';
+import Header from '../components/Header';
+import theme from '../styles/theme';
+import { NativeStackNavigationProp } from '@react-navigation/native-stack';
+import AsyncStorage from '@react-native-async-storage/async-storage';
+import { Appointment } from '../types/appointments';
+import { Doctor } from '../types/doctors';
+import { RootStackParamList } from '../types/navigation';
+import { useFocusEffect } from '@react-navigation/native';
+import { authApiService } from '../services/authApi';
+import { User } from '../types/auth';
+
+type HomeScreenProps = {
+  navigation: NativeStackNavigationProp<RootStackParamList, 'Home'>;
+};
+
+// Médicos agora vêm da API - dados removidos
+
+const HomeScreen: React.FC<HomeScreenProps> = ({ navigation }) => {
+  const [appointments, setAppointments] = useState<Appointment[]>([]);
+  const [refreshing, setRefreshing] = useState(false);
+  const [doctors, setDoctors] = useState<User[]>([]);
+
+  const loadAppointments = async () => {
+    try {
+      const storedAppointments = await AsyncStorage.getItem('appointments');
+      if (storedAppointments) {
+        setAppointments(JSON.parse(storedAppointments));
+      }
+    } catch (error) {
+      console.error('Erro ao carregar consultas:', error);
+    }
+  };
+
+  const loadDoctors = async () => {
+    try {
+      const doctorsData = await authApiService.getAllDoctors();
+      setDoctors(doctorsData);
+      console.log(`${doctorsData.length} médicos carregados no HomeScreen`);
+    } catch (error) {
+      console.error('Erro ao carregar médicos no HomeScreen:', error);
+      // Não mostra erro para o usuário no HomeScreen, apenas loga
+    }
+  };
+
+  useFocusEffect(
+    React.useCallback(() => {
+      loadAppointments();
+      loadDoctors();
+    }, [])
+  );
+
+  const onRefresh = async () => {
+    setRefreshing(true);
+    await Promise.all([loadAppointments(), loadDoctors()]);
+    setRefreshing(false);
+  };
+
+  const getDoctorInfo = (doctorId: string): User | undefined => {
+    return doctors.find(doctor => doctor.id === doctorId);
+  };
+
+  const renderAppointment = ({ item }: { item: Appointment }) => {
+    const doctor = getDoctorInfo(item.doctorId);
+    
+    return (
+      <AppointmentCard>
+        <DoctorImage source={{ uri: doctor?.image || 'https://via.placeholder.com/100' }} />
+        <InfoContainer>
+          <DoctorName>{doctor?.name || 'Médico não encontrado'}</DoctorName>
+          <DoctorSpecialty>
+            {doctor?.role === 'doctor' && 'specialty' in doctor 
+              ? doctor.specialty 
+              : 'Especialidade não encontrada'}
+          </DoctorSpecialty>
+          <DateTime>{new Date(item.date).toLocaleDateString()} - {item.time}</DateTime>
+          <Description>{item.description}</Description>
+          <Status status={item.status}>
+            {item.status === 'pending' ? 'Pendente' : 'Confirmado'}
+          </Status>
+          <ActionButtons>
+            <ActionButton>
+              <Icon name="edit" type="material" size={20} color={theme.colors.primary} />
+            </ActionButton>
+            <ActionButton>
+              <Icon name="delete" type="material" size={20} color={theme.colors.error} />
+            </ActionButton>
+          </ActionButtons>
+        </InfoContainer>
+      </AppointmentCard>
+    );
+  };
+
+  return (
+    <Container>
+      <Header />
+      <TitleContainer>
+        <Title>Minhas Consultas</Title>
+      </TitleContainer>
+
+      <Content>
+        <Button
+          title="Agendar Nova Consulta"
+          icon={
+            <FontAwesome
+              name="calendar-plus-o"
+              size={20}
+              color="white"
+              style={{ marginRight: 8 }}
+            />
+          }
+          buttonStyle={{
+            backgroundColor: theme.colors.primary,
+            borderRadius: 8,
+            padding: 12,
+            marginBottom: theme.spacing.medium
+          }}
+          onPress={() => navigation.navigate('CreateAppointment')}
+        />
+
+        <AppointmentList
+          data={appointments}
+          keyExtractor={(item: Appointment) => item.id}
+          renderItem={renderAppointment}
+          refreshControl={
+            <RefreshControl refreshing={refreshing} onRefresh={onRefresh} />
+          }
+          ListEmptyComponent={
+            <EmptyText>Nenhuma consulta agendada</EmptyText>
+          }
+        />
+      </Content>
+    </Container>
+  );
+};
+
+const Container = styled.View`
+  flex: 1;
+  background-color: ${theme.colors.background};
+`;
+
+const Content = styled.View`
+  flex: 1;
+  padding: ${theme.spacing.medium}px;
+`;
+
+const AppointmentList = styled(FlatList)`
+  flex: 1;
+`;
+
+const AppointmentCard = styled.View`
+  background-color: ${theme.colors.white};
+  border-radius: 8px;
+  padding: ${theme.spacing.medium}px;
+  margin-bottom: ${theme.spacing.medium}px;
+  flex-direction: row;
+  align-items: center;
+  elevation: 2;
+  shadow-color: #000;
+  shadow-opacity: 0.1;
+  shadow-radius: 4px;
+  shadow-offset: 0px 2px;
+`;
+
+const DoctorImage = styled.Image`
+  width: 60px;
+  height: 60px;
+  border-radius: 30px;
+  margin-right: ${theme.spacing.medium}px;
+`;
+
+const InfoContainer = styled.View`
+  flex: 1;
+`;
+
+const DoctorName = styled.Text`
+  font-size: ${theme.typography.subtitle.fontSize}px;
+  font-weight: ${theme.typography.subtitle.fontWeight};
+  color: ${theme.colors.text};
+`;
+
+const DoctorSpecialty = styled.Text`
+  font-size: ${theme.typography.body.fontSize}px;
+  color: ${theme.colors.text};
+  opacity: 0.8;
+  margin-bottom: 4px;
+`;
+
+const DateTime = styled.Text`
+  font-size: ${theme.typography.body.fontSize}px;
+  color: ${theme.colors.primary};
+  margin-top: 4px;
+`;
+
+const Description = styled.Text`
+  font-size: ${theme.typography.body.fontSize}px;
+  color: ${theme.colors.text};
+  opacity: 0.8;
+  margin-top: 4px;
+`;
+
+const Status = styled.Text<{ status: string }>`
+  font-size: ${theme.typography.body.fontSize}px;
+  color: ${(props: { status: string }) => props.status === 'pending' ? theme.colors.error : theme.colors.success};
+  margin-top: 4px;
+  font-weight: bold;
+`;
+
+const ActionButtons = styled.View`
+  flex-direction: row;
+  justify-content: flex-end;
+  margin-top: ${theme.spacing.small}px;
+`;
+
+const ActionButton = styled(TouchableOpacity)`
+  padding: ${theme.spacing.small}px;
+  margin-left: ${theme.spacing.small}px;
+`;
+
+const EmptyText = styled.Text`
+  text-align: center;
+  color: ${theme.colors.text};
+  opacity: 0.6;
+  margin-top: ${theme.spacing.large}px;
+`;
+
+const TitleContainer = styled.View`
+  padding: 16px;
+  background-color: ${theme.colors.background};
+`;
+
+const Title = styled.Text`
+  font-size: 24px;
+  font-weight: bold;
+  color: ${theme.colors.text};
+  text-align: center;
+`;
+
+export default HomeScreen;
